task03.py

+++ task8.py
@@ -1,4 +1,4 @@
-narx = int(input("1 kg konfet narxini kiriting: "))
+narx = float(input("1 kg konfet narxini kiriting: "))
 
 for kg in range(1, 11):
     print(f"{kg} kg konfet narxi: {kg * narx}")
